[共通] update

+++ Siv3D/include/Siv3D/BorderType.hpp
@@ -14,7 +14,7 @@
 
 namespace s3d
 {
-	enum class BorderType : uint32
+	enum class BorderType : uint8
 	{
 		Replicate,
