+++ Siv3D/include/Siv3D/Cursor.hpp
@@ -15,7 +15,7 @@
 
 namespace s3d
 {
-	enum class CursorStyle : uint32;
+	enum class CursorStyle : uint8;
 	struct Mat3x2;
 	class Image;
