+++ Siv3D/include/Siv3D/FontStyle.hpp
@@ -16,7 +16,7 @@
 namespace s3d
 {
 	/// @brief フォントのスタイル
-	enum class FontStyle : uint32
+	enum class FontStyle : uint8
 	{
 		/// @brief デフォルト
 		Default = 0x0,
