+++ Siv3D/include/Siv3D/PolygonFailureType.hpp
@@ -15,7 +15,7 @@
 namespace s3d
 {
 	/// @brief Polygon の入力頂点の検証結果
-	enum class PolygonFailureType : uint32
+	enum class PolygonFailureType : uint8
 	{
 		OK,
